+++ update.sh
@@ -24,24 +24,44 @@ cecho "blue" "Starting dotfiles update..."
 source "$CDIR/system-update.sh"
 
 if [[ -x "$(command -v flatpak)" ]]; then
-  flatpak update -y
+  if [ "$DRY_RUN" -ne "1" ]; then
+    flatpak update -y
+  else
+    cecho "yellow" "DRY-RUN: flatpak update -y"
+  fi
 fi
 
 if [[ -x "$(command -v brew)" ]]; then
-  brew upgrade -y
+  if [ "$DRY_RUN" -ne "1" ]; then
+    brew upgrade -y
+  else
+    cecho "yellow" "DRY-RUN: brew upgrade -y"
+  fi
 fi
 
 if [[ -x "$(command -v npm)" ]]; then
-  sudo npm install -g npm
+  if [ "$DRY_RUN" -ne "1" ]; then
+    sudo npm install -g npm
+  else
+    cecho "yellow" "DRY-RUN: sudo npm install -g npm"
+  fi
 fi
 
 if [[ -x "$(command -v uv)" ]]; then
-  uv self update
-  uv tool upgrade --all
+  if [ "$DRY_RUN" -ne "1" ]; then
+    uv self update
+    uv tool upgrade --all
+  else
+    cecho "yellow" "DRY-RUN: uv self update && uv tool upgrade --all"
+  fi
 fi
 
 if [[ -x "$(command -v oh-my-posh)" ]]; then
-  sudo oh-my-posh upgrade
+  if [ "$DRY_RUN" -ne "1" ]; then
+    sudo oh-my-posh upgrade
+  else
+    cecho "yellow" "DRY-RUN: sudo oh-my-posh upgrade"
+  fi
 fi
 
 ## End
